Added copy od released circuit

+++ trunk/arduino/projects2/barobot_carret/source/barobot_carret_main.h
@@ -12,7 +12,7 @@
 //-------------------------------------------------------------------
  
 // Put yout declarations here
- long unsigned decodeInt(String input, int odetnij );
+ long unsigned decodeInt(String input, byte odetnij );
 //-------------------------------------------------------------------
  
 //===================================================================
